+++ src/main/com/abc-bank/dao/transactions/UpdateDebit.java
@@ -8,14 +8,14 @@ public class UpdateDebit {
     private UpdateDebit() {
     } // private constructor
 
-    public static void debit(String accountNumber, long amount, String atmId, JsonObject response) {
+    public static void debit(String accountNumber, double amount, String atmId, JsonObject response) {
         // update balance
 
         try{
             Connection con = dao.connection.Connection.getConnection("transactions");
             java.sql.PreparedStatement ps = con.prepareStatement("INSERT INTO debit(account_number, debit_amount, debit_time, debit_source, debit_source_id) VALUES(?, ?, ?, ?, ?)");
             ps.setString(1, accountNumber);
-            ps.setLong(2, amount);
+            ps.setDouble(2, amount);
             ps.setTimestamp(3, model.utilities.GetTimeStamp.getTimeStamp());
             ps.setString(4, "atm");
             ps.setString(5, atmId);
