+++ src/main/com/abc-bank/dao/transactions/rollbacks/AtmDeposit.java
@@ -4,12 +4,12 @@ public class AtmDeposit {
     private AtmDeposit() {
     } // private constructor
 
-    public static void update(String accountNumber, long amount) {
+    public static void update(String accountNumber, double amount) {
         try {
             java.sql.Connection con = dao.connection.Connection.getConnection("transactions");
             java.sql.PreparedStatement ps = con.prepareStatement("INSERT INTO atm_deposit_rollbacks (account_number, amount, t_time) VALUES (?, ?, ?)");
             ps.setString(1, accountNumber);
-            ps.setLong(2, amount);
+            ps.setDouble(2, amount);
             ps.setTimestamp(3, model.utilities.GetTimeStamp.getTimeStamp());
             ps.executeUpdate();
         }catch(Exception e){
