+++ src/main/com/abc-bank/dao/transactions/rollbacks/AtmWithdraw.java
@@ -8,12 +8,12 @@ public class AtmWithdraw {
     private AtmWithdraw() {
     } // private constructor
 
-    public static void update(String accountNumber, long amount, JsonObject response) {
+    public static void update(String accountNumber, double amount, JsonObject response) {
         try{
             Connection con = dao.connection.Connection.getConnection("transactions");
             java.sql.PreparedStatement ps = con.prepareStatement("INSERT INTO atm_withdraw_rollbacks (account_number, amount, t_time) VALUES (?, ?, ?)");
             ps.setString(1, accountNumber);
-            ps.setLong(2, amount);
+            ps.setDouble(2, amount);
             ps.setTimestamp(3, model.utilities.GetTimeStamp.getTimeStamp());
             ps.executeUpdate();
         }catch(Exception e){
