+++ src/main/com/abc-bank/dao/transactions/rollbacks/NetbankingTransaction.java
@@ -9,12 +9,12 @@ public class NetbankingTransaction {
     } // private constructor
 
 
-    public static void rollbackFrom(String from_account, long amount) {
+    public static void rollbackFrom(String from_account, double amount) {
         try{
             Connection con = dao.connection.Connection.getConnection("transactions");
             java.sql.PreparedStatement ps = con.prepareStatement("insert into netbanking_withdraw_rollbacks (account_number, amount, t_time) values (?, ?, ?)");
             ps.setString(1, from_account);
-            ps.setLong(2, amount);
+            ps.setDouble(2, amount);
             ps.setTimestamp(3, GetTimeStamp.getTimeStamp());
             ps.executeUpdate();
         }catch(Exception e){
@@ -22,12 +22,12 @@ public class NetbankingTransaction {
         }
     }
 
-    public static void rollbackTo(String to_account, long amount) {
+    public static void rollbackTo(String to_account, double amount) {
         try{
             Connection con = dao.connection.Connection.getConnection("transactions");
             java.sql.PreparedStatement ps = con.prepareStatement("insert into netbanking_deposit_rollbacks (account_number, amount, t_time) values (?, ?, ?)");
             ps.setString(1, to_account);
-            ps.setLong(2, amount);
+            ps.setDouble(2, amount);
             ps.setTimestamp(3, GetTimeStamp.getTimeStamp());
             ps.executeUpdate();
         }catch(Exception e){
